Move Scara port mocks in separate file
