+++ master-firmware/tests/scara_kinematics.cpp
@@ -17,7 +17,7 @@ TEST(GetAngleFromArmPosTestGroup, SimpleShoulderAngle)
     point_t elbow, hand;
     elbow.x = 100;
     elbow.y = 0;
-    angle = compute_shoulder_angle(elbow, hand);
+    angle = scara_compute_shoulder_angle(elbow, hand);
 
     DOUBLES_EQUAL(RAD(0), angle, 1e-3);
 }
@@ -28,6 +28,6 @@ TEST(GetAngleFromArmPosTestGroup, SimpleElbowAngle)
     point_t elbow = {10, 10};
     point_t hand =  {20, 20};
 
-    angle = compute_elbow_angle(elbow, hand);
+    angle = scara_compute_elbow_angle(elbow, hand);
     DOUBLES_EQUAL(RAD(45), angle, 1e-3);
 }
